pyproject.toml
[project]
name = "pixel-rush"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "numpy>=2.0.2",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "streamlit>=1.22.0",
]
